+++ handlers/json.js
@@ -3,13 +3,13 @@
   module.exports = function(entry) {
     var data, hashtags, re, tag;
 
-    tag = entry.match(/\w+/);
+    tag = entry.match(/[\w-]+/);
     tag = tag[0];
     entry = '{' + entry.substr(tag.length).trim() + '}';
-    re = /\:\s*(?![."\s\d])([^{}[\]",]+)\s*(?=[,}]|$)/g;
+    re = /\:\s*(?![."\s])((?:[^{}[\]",]+)\s*)(?=[,}]|$)/g;
     entry = entry.replace(re, ':"$1"');
     hashtags = [];
-    entry = entry.replace(/([{,]\s*)(\w+)\:/g, function(whole, $1, $2) {
+    entry = entry.replace(/([{,]\s*)([\w-]+)\:/g, function(whole, $1, $2) {
       return $1 + '"' + $2 + '":';
     });
     re = /#(\w+)\s*([,}]|$)/g;