+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.0.5
+
+* Display hashtags
+* More intuitive data entry
+* Ability to delete snippets
+
 ### 0.0.4
 
 * Added hashtag support